+++ scripts/common/constants.ts
