+++ scripts/genIndex.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 const fs = require('fs');
 
-const genExports = require('./common/genExports');
-const genVersion = require('./common/genVersion');
-const { SRC_PATH } = require('./common/constants');
+const genExports = require('./common/genExports.ts');
+const genVersion = require('./common/genVersion.ts');
+const { SRC_PATH } = require('./common/constants.ts');
 
 fs.writeFileSync(SRC_PATH + '/index.tsx', Buffer.from(`${genVersion()};\r\n\r\n${genExports()}`), (err) => {
   if (err) {
