+++ scripts/serve.ts
